Refactor and clean up various components
- Removed unused variables and simplified code in Vulkan backend files.
- Updated model registration entries in BERT and ViT to include missing SHA256 fields.
- Suppressed specific compiler warnings around deprecated constructors in SyncBatchNorm.
- Improved handling of output shapes in Graph's infer_types and infer_symbolic_types methods.
- Enhanced test cases to intentionally exercise deprecated constructors for backward compatibility.
- Cleaned up unnecessary lambda functions and improved readability in RNN utilities.

+++ src/backends/cpu/kernels/fused_ops.cpp
@@ -97,7 +97,7 @@ auto fused_linear_relu_kernel(
                 float* out_data = output.data<float>();
                 const float* bias_data = b_ptr ? b_ptr->data<float>() : nullptr;
 
-                #pragma omp parallel for if(total_elems > ::tenzor::OmpThresholds::simple())
+                #pragma omp parallel for if(static_cast<int64_t>(total_elems) > ::tenzor::OmpThresholds::simple())
                 for (size_t i = 0; i < static_cast<size_t>(batch_size); ++i) {
                     for (size_t j = 0; j < static_cast<size_t>(out_features); ++j) {
                         size_t idx = i * out_features + j;
@@ -112,7 +112,7 @@ auto fused_linear_relu_kernel(
                 double* out_data = output.data<double>();
                 const double* bias_data = b_ptr ? b_ptr->data<double>() : nullptr;
 
-                #pragma omp parallel for if(total_elems > ::tenzor::OmpThresholds::simple())
+                #pragma omp parallel for if(static_cast<int64_t>(total_elems) > ::tenzor::OmpThresholds::simple())
                 for (size_t i = 0; i < static_cast<size_t>(batch_size); ++i) {
                     for (size_t j = 0; j < static_cast<size_t>(out_features); ++j) {
                         size_t idx = i * out_features + j;
@@ -172,13 +172,13 @@ auto fused_conv2d_relu_kernel(
             int64_t n = result.numel();
             if (result.dtype() == DType::Float32) {
                 float* data = result.data<float>();
-                #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+                #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
                 for (int64_t i = 0; i < n; ++i) {
                     data[i] = std::max(0.0f, data[i]);
                 }
             } else if (result.dtype() == DType::Float64) {
                 double* data = result.data<double>();
-                #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+                #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
                 for (int64_t i = 0; i < n; ++i) {
                     data[i] = std::max(0.0, data[i]);
                 }
@@ -214,13 +214,13 @@ auto fused_conv2d_relu_kernel(
             int64_t n = result.numel();
             if (result.dtype() == DType::Float32) {
                 float* data = result.data<float>();
-                #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+                #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
                 for (int64_t i = 0; i < n; ++i) {
                     data[i] = std::max(0.0f, data[i]);
                 }
             } else if (result.dtype() == DType::Float64) {
                 double* data = result.data<double>();
-                #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+                #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
                 for (int64_t i = 0; i < n; ++i) {
                     data[i] = std::max(0.0, data[i]);
                 }
@@ -613,7 +613,7 @@ auto fused_add_relu_kernel(const Tensor& a, const Tensor& b) -> Tensor {
 #if defined(TENZOR_ATTN_AVX2)
         size_t vec_end = n - (n % 8);
         __m256 zero_vec = _mm256_setzero_ps();
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (size_t i = 0; i < vec_end; i += 8) {
             __m256 v = _mm256_loadu_ps(data + i);
             _mm256_storeu_ps(data + i, _mm256_max_ps(zero_vec, v));
@@ -622,7 +622,7 @@ auto fused_add_relu_kernel(const Tensor& a, const Tensor& b) -> Tensor {
             data[i] = std::max(0.0f, data[i]);
         }
 #else
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (size_t i = 0; i < n; ++i) {
             data[i] = std::max(0.0f, data[i]);
         }
@@ -630,7 +630,7 @@ auto fused_add_relu_kernel(const Tensor& a, const Tensor& b) -> Tensor {
     } else if (result.dtype() == DType::Float64) {
         double* data = result.data<double>();
         size_t n = static_cast<size_t>(result.numel());
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (size_t i = 0; i < n; ++i) {
             data[i] = std::max(0.0, data[i]);
         }
@@ -638,7 +638,7 @@ auto fused_add_relu_kernel(const Tensor& a, const Tensor& b) -> Tensor {
         Tensor result_f32 = result.to(DType::Float32);
         float* data = result_f32.data<float>();
         size_t n = static_cast<size_t>(result.numel());
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (size_t i = 0; i < n; ++i) {
             data[i] = std::max(0.0f, data[i]);
         }
@@ -647,7 +647,7 @@ auto fused_add_relu_kernel(const Tensor& a, const Tensor& b) -> Tensor {
         Tensor result_f32 = result.to(DType::Float32);
         float* data = result_f32.data<float>();
         size_t n = static_cast<size_t>(result.numel());
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (size_t i = 0; i < n; ++i) {
             data[i] = std::max(0.0f, data[i]);
         }
@@ -686,7 +686,7 @@ auto fused_gelu_kernel(const Tensor& input) -> Tensor {
         float* out_data = result.data<float>();
         size_t n = static_cast<size_t>(input.numel());
 
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (size_t i = 0; i < n; ++i) {
             float x = in_data[i];
             out_data[i] = 0.5f * x * (1.0f + std::erf(x * INV_SQRT2_F));
@@ -696,7 +696,7 @@ auto fused_gelu_kernel(const Tensor& input) -> Tensor {
         double* out_data = result.data<double>();
         size_t n = static_cast<size_t>(input.numel());
 
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (size_t i = 0; i < n; ++i) {
             double x = in_data[i];
             out_data[i] = 0.5 * x * (1.0 + std::erf(x * INV_SQRT2_D));
@@ -706,7 +706,7 @@ auto fused_gelu_kernel(const Tensor& input) -> Tensor {
         Float16* out_data = result.data<Float16>();
         size_t n = static_cast<size_t>(input.numel());
 
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (size_t i = 0; i < n; ++i) {
             float x = static_cast<float>(in_data[i]);
             out_data[i] = Float16(0.5f * x * (1.0f + std::erf(x * INV_SQRT2_F)));
@@ -716,7 +716,7 @@ auto fused_gelu_kernel(const Tensor& input) -> Tensor {
         BFloat16* out_data = result.data<BFloat16>();
         size_t n = static_cast<size_t>(input.numel());
 
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (size_t i = 0; i < n; ++i) {
             float x = static_cast<float>(in_data[i]);
             out_data[i] = BFloat16(0.5f * x * (1.0f + std::erf(x * INV_SQRT2_F)));
@@ -1961,7 +1961,7 @@ auto fused_sgd_step_kernel(Tensor& param, const Tensor& grad, Tensor* momentum_b
         float* p = param.data<float>();
         const float* g = grad.data<float>();
 
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (int64_t i = 0; i < n; ++i) {
             float grad_val = g[i];
             if (weight_decay != 0.0f) {
@@ -1984,7 +1984,7 @@ auto fused_sgd_step_kernel(Tensor& param, const Tensor& grad, Tensor* momentum_b
         double* p = param.data<double>();
         const double* g = grad.data<double>();
 
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (int64_t i = 0; i < n; ++i) {
             double grad_val = g[i];
             if (weight_decay != 0.0) {
@@ -2031,7 +2031,7 @@ auto fused_adam_step_kernel(Tensor& param, const Tensor& grad,
         float* v = exp_avg_sq.data<float>();
         float* v_max = (amsgrad && max_exp_avg_sq) ? max_exp_avg_sq->data<float>() : nullptr;
 
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (int64_t i = 0; i < n; ++i) {
             float grad_val = g[i];
 
@@ -2062,7 +2062,7 @@ auto fused_adam_step_kernel(Tensor& param, const Tensor& grad,
         double* v = exp_avg_sq.data<double>();
         double* v_max = (amsgrad && max_exp_avg_sq) ? max_exp_avg_sq->data<double>() : nullptr;
 
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (int64_t i = 0; i < n; ++i) {
             double grad_val = g[i];
             if (decoupled_weight_decay && weight_decay != 0.0) {
@@ -2110,7 +2110,7 @@ auto fused_adam_atan2_step_kernel(Tensor& param, const Tensor& grad,
         float* v = exp_avg_sq.data<float>();
         float* v_max = (amsgrad && max_exp_avg_sq) ? max_exp_avg_sq->data<float>() : nullptr;
 
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (int64_t i = 0; i < n; ++i) {
             float grad_val = g[i];
             if (weight_decay != 0.0f) {
@@ -2141,7 +2141,7 @@ auto fused_adam_atan2_step_kernel(Tensor& param, const Tensor& grad,
         double d_bias_correction2 = 1.0 - std::pow(static_cast<double>(beta2), static_cast<double>(step));
         double d_step_size = static_cast<double>(lr) / (1.0 - std::pow(static_cast<double>(beta1), static_cast<double>(step)));
 
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (int64_t i = 0; i < n; ++i) {
             double grad_val = g[i];
             if (weight_decay != 0.0f) {
@@ -2185,7 +2185,7 @@ auto fused_rmsprop_step_kernel(Tensor& param, const Tensor& grad,
         float* ga = (centered && grad_avg) ? grad_avg->data<float>() : nullptr;
         float* buf = (momentum > 0.0f && momentum_buffer) ? momentum_buffer->data<float>() : nullptr;
 
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (int64_t i = 0; i < n; ++i) {
             float grad_val = g[i];
             if (weight_decay != 0.0f) {
@@ -2216,7 +2216,7 @@ auto fused_rmsprop_step_kernel(Tensor& param, const Tensor& grad,
         double* ga = (centered && grad_avg) ? grad_avg->data<double>() : nullptr;
         double* buf = (momentum > 0.0f && momentum_buffer) ? momentum_buffer->data<double>() : nullptr;
 
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (int64_t i = 0; i < n; ++i) {
             double grad_val = g[i];
             if (weight_decay != 0.0f) {
@@ -2261,7 +2261,7 @@ auto fused_adadelta_step_kernel(Tensor& param, const Tensor& grad,
         float* sq = square_avg.data<float>();
         float* ad = acc_delta.data<float>();
 
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (int64_t i = 0; i < n; ++i) {
             float grad_val = g[i];
             if (weight_decay != 0.0f) {
@@ -2279,7 +2279,7 @@ auto fused_adadelta_step_kernel(Tensor& param, const Tensor& grad,
         double* sq = square_avg.data<double>();
         double* ad = acc_delta.data<double>();
 
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (int64_t i = 0; i < n; ++i) {
             double grad_val = g[i];
             if (weight_decay != 0.0f) {
@@ -2312,7 +2312,7 @@ auto fused_adagrad_step_kernel(Tensor& param, const Tensor& grad,
         const float* g = grad.data<float>();
         float* ss = sum_sq.data<float>();
 
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (int64_t i = 0; i < n; ++i) {
             float grad_val = g[i];
             if (weight_decay != 0.0f) {
@@ -2327,7 +2327,7 @@ auto fused_adagrad_step_kernel(Tensor& param, const Tensor& grad,
         double* ss = sum_sq.data<double>();
         double clr_d = static_cast<double>(lr) / (1.0 + static_cast<double>(step - 1) * static_cast<double>(lr_decay));
 
-        #pragma omp parallel for if(n > ::tenzor::OmpThresholds::simple())
+        #pragma omp parallel for if(static_cast<int64_t>(n) > ::tenzor::OmpThresholds::simple())
         for (int64_t i = 0; i < n; ++i) {
             double grad_val = g[i];
             if (weight_decay != 0.0f) {
